requirements.txt
flask
requests
python-dotenv
google-genai
